The module 'clients' was created. In this module u deligate REST API endpoints which are exposed by third-party or microservice.

+++ fraud/src/main/java/org/dimasik666/fraud/domain/model/FraudCheckHistory.java
@@ -23,8 +23,8 @@ public class FraudCheckHistory {
             strategy = GenerationType.SEQUENCE,
             generator = "fraud_id_sequence"
     )
-    private Integer id;
-    private Integer customerId;
+    private long id;
+    private long customerId;
     private Boolean isFraudster;
     private LocalDateTime createdAt;
