+++ fraud/src/main/java/org/dimasik666/fraud/service/FraudCheckService.java
@@ -1,5 +1,5 @@
 package org.dimasik666.fraud.service;
 
 public interface FraudCheckService {
-    boolean isFraudulentCustomer(Integer customerId);
+    boolean isFraudulentCustomer(long customerId);
 }
